+++ packages/plugin-htmlparser2/test/pick1.ts
@@ -1,20 +1,20 @@
-import test, {ExecutionContext} from 'ava';
+import test, { type ExecutionContext } from 'ava';
 import * as htmlparser2 from 'htmlparser2';
 import { DecisionTree } from 'selderee';
 
-import { hp2Builder } from '../src/hp2-builder';
+import { hp2Builder } from '../src/hp2-builder.ts';
 
-const html = /*html*/`<html><body>
+const html = /* html */`<html><body>
   <div><p id="A" class="foo qux">second</p></div>
 </body></html>`;
 const dom = htmlparser2.parseDocument(html);
 
-function pick1Macro(
+function pick1Macro (
   t: ExecutionContext,
   elemId: string,
   selectors: [string, string][],
   preferFirst: boolean,
-  expected: string | null
+  expected: string | null,
 ) {
   const dt = (new DecisionTree(selectors)).build(hp2Builder);
   const element = htmlparser2.DomUtils.getElementById(elemId, dom.children, true);
@@ -23,7 +23,7 @@ function pick1Macro(
   } else {
     t.is(
       dt.pick1(element, preferFirst),
-      expected
+      expected,
     );
   }
 }
@@ -38,8 +38,8 @@ const selectors: [string, string][] = [
   ['p.bar', 'A3'],
   ['p.foo', 'A4'],
   ['div > [class~=qux]', 'A5'],
-  ['p[class]','A6'],
-  ['*.baz', 'A7']
+  ['p[class]', 'A6'],
+  ['*.baz', 'A7'],
 ];
 
 test('multiple inputs, prefer last', pick1Macro, 'A', selectors, false, 'A6');