+++ packages/plugin-htmlparser2/test/pickAll.ts
@@ -1,18 +1,18 @@
-import test, {ExecutionContext} from 'ava';
+import test, { type ExecutionContext } from 'ava';
 import * as htmlparser2 from 'htmlparser2';
 import { DecisionTree } from 'selderee';
 
-import { hp2Builder } from '../src/hp2-builder';
+import { hp2Builder } from '../src/hp2-builder.ts';
 
-const html = /*html*/`<html><body>
+const html = /* html */`<html><body>
   <div><p id="A" class="foo qux">second</p></div>
 </body></html>`;
 const dom = htmlparser2.parseDocument(html);
 
-function pickAllMacro(
+function pickAllMacro (
   t: ExecutionContext,
   elemId: string,
-  selectors: [string, string][]
+  selectors: [string, string][],
 ) {
   const dt = (new DecisionTree(selectors)).build(hp2Builder);
   const element = htmlparser2.DomUtils.getElementById(elemId, dom.children, true);
@@ -21,7 +21,7 @@ function pickAllMacro(
   } else {
     t.snapshot(
       dt.pickAll(element),
-      `For \`#${elemId}\`, pick matches from \`${JSON.stringify(selectors)}\``
+      `For \`#${elemId}\`, pick matches from \`${JSON.stringify(selectors)}\``,
     );
   }
 }
@@ -37,5 +37,5 @@ test('multiple inputs', pickAllMacro, 'A', [
   ['p.bar', 'A4'],
   ['*.baz', 'A5'],
   ['* > [class~=qux]', 'A6'],
-  ['#A[class]','A7']
+  ['#A[class]', 'A7'],
 ]);