+++ packages/plugin-htmlparser2/test/selectors.ts
@@ -1,19 +1,19 @@
-import test, {ExecutionContext} from 'ava';
+import test, { type ExecutionContext } from 'ava';
 import * as htmlparser2 from 'htmlparser2';
 import { DecisionTree } from 'selderee';
 
-import { hp2Builder } from '../src/hp2-builder';
+import { hp2Builder } from '../src/hp2-builder.ts';
 
-const html = /*html*/`<html><body>
+const html = /* html */`<html><body>
   <div><p id="A" class="foo qux .">second</p><div id="B" baz>second</div></div>
 </body></html>`;
 const dom = htmlparser2.parseDocument(html);
 
-function selectorsMacro(
+function selectorsMacro (
   t: ExecutionContext,
   elemId: string,
   selectors: string[],
-  isMatching: boolean
+  isMatching: boolean,
 ) {
   const dt = (new DecisionTree(selectors.map(s => ([s, s])))).build(hp2Builder);
   const element = htmlparser2.DomUtils.getElementById(elemId, dom.children, true);
@@ -22,7 +22,7 @@ function selectorsMacro(
   } else {
     t.is(
       dt.pickAll(element).length,
-      (isMatching) ? selectors.length : 0
+      (isMatching) ? selectors.length : 0,
     );
   }
 }
@@ -42,14 +42,14 @@ test('class match - escape sequence', selectorsMacro, 'A', ['.\\.'], true);
 test('attribute presence match 1', selectorsMacro, 'A', [
   '[class]',
   '[id]',
-  '[class][id]'
+  '[class][id]',
 ], true);
 
 test('attribute presence match 2', selectorsMacro, 'B', ['[baz]'], true);
 
 test('attribute presence non-match 1', selectorsMacro, 'A', [
   '[baz]',
-  '[class][baz]'
+  '[class][baz]',
 ], false);
 
 test('attribute presence non-match 2', selectorsMacro, 'B', ['[class]'], false);
@@ -60,7 +60,7 @@ test('attribute value match', selectorsMacro, 'A', [
   '[class^="foO"i]',
   '[class*="o q"]',
   '[id$=A][id|=A]',
-  '#A'
+  '#A',
 ], true);
 
 test('attribute value non-match', selectorsMacro, 'B', [
@@ -70,7 +70,7 @@ test('attribute value non-match', selectorsMacro, 'B', [
   '[class*="o q"]',
   '[id$=A]',
   '[id|=A]',
-  '#A'
+  '#A',
 ], false);
 
 test('combinators match', selectorsMacro, 'B', [
@@ -78,11 +78,11 @@ test('combinators match', selectorsMacro, 'B', [
   '* > #B[baz]',
   'div > *',
   'p + div',
-  'div > [class] + div'
+  'div > [class] + div',
 ], true);
 
 test('combinators non-match', selectorsMacro, 'A', [
   '* > #B',
   'p > *',
-  '* + *'
+  '* + *',
 ], false);