+++ packages/selderee/rollup.config.js
@@ -1,21 +1,42 @@
 import typescript from '@rollup/plugin-typescript';
 import cleanup from 'rollup-plugin-cleanup';
+import del from 'rollup-plugin-delete';
+import { dts } from 'rollup-plugin-dts';
 
 export default [
   {
     external: [ 'parseley' ],
     input: 'src/selderee.ts',
-    plugins: [ typescript(), cleanup({ extensions: ['ts'] }) ],
+    treeshake: false,
+    plugins: [
+      typescript({ compilerOptions: { declaration: false, isolatedDeclarations: false } }),
+      cleanup({ extensions: ['ts'] }),
+    ],
     output: [
       {
-        dir: 'lib',
         format: 'es',
-        entryFileNames: '[name].mjs',
+        file: 'lib/selderee.mjs',
       },
       {
-        dir: 'lib',
         format: 'cjs',
-        entryFileNames: '[name].cjs',
+        file: 'lib/selderee.cjs',
+      },
+    ],
+  },
+  {
+    input: 'src/selderee.ts',
+    plugins: [
+      dts(),
+      del({ targets: 'lib/*.d.ts', hook: 'writeBundle', verbose: true }),
+    ],
+    output: [
+      {
+        format: 'es',
+        file: 'lib/selderee.d.mts',
+      },
+      {
+        format: 'cjs',
+        file: 'lib/selderee.d.cts',
       },
     ],
   },