+++ packages/selderee/src/selderee.ts
@@ -1,5 +1,11 @@
-export * as Ast from './Ast';
-export * as Types from './Types';
-export * as Treeify from './TreeifyBuilder';
-export * from './DecisionTree';
-export * from './Picker';
+/**
+ * Selector-based decision tree - core module.
+ *
+ * @packageDocumentation
+ */
+
+export * as Ast from './Ast.ts';
+export * as Types from './Types.ts';
+export * as Treeify from './TreeifyBuilder.ts';
+export * from './DecisionTree.ts';
+export * from './Picker.ts';